LWDEV-6992 Switch to BigDecimal (#108)
* LWDEV-6992 use big decimal

* LWDEV-6992 fees use big decimal

* LWDEV-6992 matching engine use big decimal

* LWDEV-6992 matching engine use big decimal

* LWDEV-6992 use big decimal in services

* LWDEV-6992 use big decimal in services

* LWDEV-6992 use big decimal in services

* LWDEV-6992 use big decimal in tests

* LWDEV-6992 use bigdecimal in limitorderservice test

* LWDEV-6992 use bigdecimal in rounding test

* LWDEV-6992 use bigdecimal in stop limit test

* LWDEV-6992 use bigdecimal in tests

* LWDEV-6992 use bigdecimal in multilimitorder service test

* LWDEV-6992 use bigdecimal in limit service test

* LWDEV-6992 use bigdecimal market order and other tests

* LWDEV-6992 all build errors are fixed

* LWDEV-6992 fix limitorder service

* LWDEV-6992 fix matching engine divide operation prevision with bigDecimal

* LWDEV-6992 fixes

* LWDEV-6992 fix market order tests

* LWDEV-6992 fix market order tests

* LWDEV-6992 fix market order tests

* LWDEV-6992 fix tests

* LWDEV-6992 fix tests

* LWDEV-6992 fix tests

* LWDEV-6992 implement big decimal serializer

* LWDEV-6992 use setscale instead of parseDouble

* LWDEV-6992 use big math lib for big decimal operations

* LWDEV-6992 implement custom equals hashcode FeeTransfer

* LWDEV-6992 implement equals hashcode

* LWDEV-6992 implement equals hash code

* LWDEV-6992 remove not needed code

* LWDEV-6992 fix import

* LWDEV-6992 fix imports

* LWDEV-6992 merge fixes

* LWDEV-6992 fix perfromance tests

* LWDEV-6992 fix merge

* LWDEV-6992 fix merge

* LWDEV-6992 fix merge

* LWDEV-6992 use bean for stop order file db accessor

* LWDEV-6992 AbstractFileOrderBookDatabaseAccessor use nio for handling dirs with many files

* LWDEV-6992 deduplicate file name creation

* LWDEV-7668 move back to limit order

* LWDEV-7668 use limit order everywhere

* LWDEV-6992 read old format of orders from order book

* fix nulls in asset pair:

* LWDEV-6992 fix nulls in asset pair

* LWDEV-6992 merge fixes

* LWDEV-6992 refactoring

* LWDEV-6992 review fixes

* LWDEV-6992 review fixes

* LWDEV-6992 review fixes

* LWDEV-6992 create new fee instructions

* LWDEV-6992 introduce new fee instructions

* LWDEV-6992 add convertor for fee instructions

* LWDEV-6992 fix logging

+++ src/main/kotlin/com/lykke/matching/engine/daos/BestPrice.kt
@@ -1,8 +1,9 @@
 package com.lykke.matching.engine.daos
 
+import java.math.BigDecimal
 import java.util.Date
 
-class BestPrice(val asset: String, val ask: Double, val bid: Double) {
+class BestPrice(val asset: String, val ask: BigDecimal, val bid: BigDecimal) {
     var dateTime = Date()
 
     override fun toString(): String {
