+++ src/main/kotlin/com/lykke/matching/engine/outgoing/messages/CashTransferOperation.kt
@@ -1,8 +1,9 @@
 package com.lykke.matching.engine.outgoing.messages
 
-import com.lykke.matching.engine.daos.FeeInstruction
+import com.lykke.matching.engine.daos.v2.FeeInstruction
 import com.lykke.matching.engine.daos.FeeTransfer
-import com.lykke.matching.engine.daos.fee.Fee
+import com.lykke.matching.engine.daos.fee.v2.Fee
+import java.math.BigDecimal
 import java.util.Date
 
 class CashTransferOperation(
@@ -11,7 +12,7 @@ class CashTransferOperation(
         val toClientId: String,
         val dateTime: Date,
         val volume: String,
-        val overdraftLimit: Double?,
+        val overdraftLimit: BigDecimal?,
         var asset: String,
         val feeInstruction: FeeInstruction?,
         val feeTransfer: FeeTransfer?,
